requirements.txt
mysql-connector-python
pyserial